requirements.txt
pyTeleBotAPI
requests